refactor: мелкий рефактор размещения тостов
