+++ src/entities/user/lib/positionNotifications.ts
@@ -1,5 +1,5 @@
 import { errorPositionMessage, infoPositionMessage, successPositionMessage } from './constants';
-import { createToast } from '~/shared/lib';
+import createToast from '~/shared/lib/toast/createToast';
 
 export const errorPositionToast = () => createToast('error', errorPositionMessage);
 export const infoPositionToast = () => createToast('info', infoPositionMessage);
