--- src/shared/lib/index.ts
@@ -1,3 +1,0 @@
-import createToast from './utils/createToast';
-
-export { createToast };
